requirements.txt
psutil>=5.9
pystray>=0.19
Pillow>=10.0
screen-brightness-control>=0.23
pyinstaller>=6.0
